Backend refactor Part 1
Updated most database tables to use the new ERD designs
Maps have Tracks have Zones, all 3 have stats, zones have properties and geometries
Map uploading now takes the .zon file to determine track count, difficulty is still manually set per track
TODO (for part 2): expose zone geometries via endpoint for game to fetch and construct them, fix run submission and display

+++ server/src/models/db/map.js
@@ -9,7 +9,7 @@ module.exports = (sequelize, type) => {
 		},
 		name: type.STRING(32),
 		type: {
-			type: type.SMALLINT.UNSIGNED,
+			type: type.TINYINT.UNSIGNED,
 			defaultValue: 0,
 		},
 		statusFlag: {
